requirements.txt
aws-cdk.core==1.128.0
aws_cdk.aws_cloudwatch==1.128.0
aws_cdk.aws_cloudwatch_actions==1.128.0
aws-cdk.aws_iam==1.128.0
aws-cdk.aws_lambda==1.128.0
aws-cdk.aws_sns==1.128.0
aws-cdk.aws_sns_subscriptions==1.128.0
boto3==1.19.6
pyyaml<5.5, >=3.10